Add withWorker component
